rename file driver extension
